requirements.txt
PyPDF2
ollama
python-docx
groq
streamlit